ipcjs: add React co-project, migrate files
This commit refactors the `ipcjs` project by migrating its toy React app
into a project of its own and upgrading it to be able to showcase the
way the ipcjs library can be used. The React app shows how tables,
dictionaries and other types can be rendered by default in the browser,
and implements a simple text-area into which `q` commands can be typed
and then sent for evaluation on the server.