log_filter: fix bug in table scanning code
This commit updates the `src/log_filter.c` script to fix a bug in the
code which determines the length of a table object, which omitted the
attribute byte between the 0x62 and 0x63 table and dict marker-bytes
respectively. It also improves the table-name matching check, which did
not include the null byte and thus for a requirement of "trade" would
also match "trade_anything".