ipc++: bufixes and ergonomics
This commit updates the root `CMakeLists.txt` file to increase the C++
standard from C++20 to C++23. It updates the `src/CMakeLists.txt` to
include the `-fPIC` switch ... how did I forget that? It modifies the
`KdbIpcMessageWriter` constructor and class so that it no longer takes a
shared-pointer, just a `const` reference to a `KdbBase` instance. The
shared-pointer just does not suit stack-allocation for messages the user
wants to write. This commit adds more `format_to` definitions so some of
the `enum class` types now print themselves usefully.