ipc++,tmpux: add KdbJournalReader & filter_msgs impl
This commit adds some as yet untested functions to `KdbType.h` and its
`.cpp` file to scan a journal's individual messages to ascertain their
length, and the `filter_msgs` function in the kdb-subscribe CPP file in
the TPMux project. The functions work for this use-case: they copy as
efficiently as possible the messages matching a given set of table-names
from the source TP journal to the destination journal.