tpmux: start fix the co_return/no continuation issue
This commit tries to address the problem I've got with my suspend-never
Task implementation that tries to co_return -1 before it's been awaited.
One way is to detect the null continuation and raise an error, the other
is simply to raise an error instead of returning -1. I'm not quite sure
what the micro-differences of this approach are. This came to light when
I began writing tests for it. The success-case, of course, was fine.

I'm going to do some work on a parallel implementation that doesn't
daisy-chain the coroutines, but rather, has a top-level orchestrator
that passes on the socket FD, so more of a hub-and-spoke approach.
Perhaps I might even create something nice that co_yields messages,
transparently either from a TP journal or its TCP stream. Then the
filtering and the rest of it would all be the same.