tpmux, ipc++: use clang, add KdbUpdMsgFilter
Some refactoring as well as housekeeping. After an upgrade to Fedora-42
and GCC+15, the ipc++ project and the formatters in KdbType no longer
compile. See `ipc++/configure.sh` for the error and some of the details
that I found online about this particular issue. I've no idea why it
manifests, but it compiles with clang.

I've changed the names of a `KdbJournalReader` to `KdbJnlMsgFilter` and
created `KdbUpdMsgFilter`. As yet untested ...