tpmux: initial project add
This commit adds the files which comprise the initial "working" (ho ho)
cut of the "TP Mux" coroutine project. The project is intended to be a
low-overhead filtering proxy for multiple tickerplant streams, with the
intention that it writes its own unified log and efficiently forwards
messages matching the filter criteria to its own subscribers.