ipc++: fixed bug in WriteBuf::writeAry, add CircularBuffer
This commit fixes a nasty little issue affecting the
`WriteBuf::writeAry` function, which was causing it to write data from
an incorrect byte-offset to the destination-buffer.

This commit addsd the `CircularBuffer` `.h` and `.cpp` files, which
implement a "magic" wrap-around buffer that obviates the need to compact
data when you end up with a rump few bytes right in the middle. Check it
out, it's pretty cool.

This commit adds the ultra-basic test `test/CircularBufferTest.cpp`, but
probably more importantly, adds the example
`examples/demo_circular_buffer.cpp`.