ipc++: add examples/subscribe_to_tp, bug-fix filter_msg
This commit adds a new application to the examples directory which
demonstrates how to subscribe to a tickerplant, replay its journal
and then listen for `up` messages for the `trade` table.

This commit also fixes some really nasty bugs in the `_filter_upd_msg`
function, which ought to have defeated its purpose. Sorry about that.

This commit updates a few of the README files.