ipcjs/react: code tidy, bugfix, add react/README.md
This commit tidies a small corner of the JS in `ipcjs/src/ipc.js` in
preparation for using the JS IPC library in a forthcoming project which
(at the moment) is loading it as an ES6 module.

This commit makes testing the React integration easier by documenting
how to build and run it ... which was quite the oversight.

This commit also moves its port from 30098 to 30097 since I've been
using that for one of the `tpmux` project's tickerplants