ipc++: add CircularBufferTest, see description
This commit renames the previous `CircularBufferTest` to
`CircularBufferIntegrationTest`, which it is, and reimplements the
former as a unit-test which leans on GMock matchers for the IO
functions. This adds the new file `test/KdbIoMockDefs.h`.

This commit modifies the root `CMakeLists.txt` file slightly to hoist
some of the configuration out of the script, as I've recently seen
suggested in a conference video. These are now set (for me, at least) in
my environment. A comment makes this clear to readers.