ipc++: add test coverage for CircularBuffer.cpp
This commit adds further and fairly complete coverate for the
`init_circ_buffer` function in `CircularBuffer.cpp`, and fixes a few
bugs in the error handling, _e.g._ calling <expected>.error() when, in
fact, the call succeeded but the subsequent equality test failed. There
was also an extraneous `%s` instead of a `{}` in there. Oops.