krb5: have got it working, includes cookie code
The commit includes some commented-out cookie-setting and storing code;
it turns out that what was needed was changing the hostname on the
websocket itself, from `localhost` to `window.location.hostname`.

The .h.ka keep-alive feature seems to help, although that won't work
with an older verison of q. It might just be easier to go for
stateless/Connection: close semantics, Firefox seems to manage ok.