krb5: have mutual-auth working; changed web pathway
I've changed things quite a bit in this commit. The context is no longer
allowed to persist in memory between calls, and instead it's "exported"
between accept/init ctx calls. This is nice because we've got nothing
left hanging around, and nice because instead of a live pointer hanging
around, we've now just got around 2k base64-encoded characters instead,
which is easy to clean up by removing it (at some point) from the table.

I've sort-of separated the workflows for HTTP (now in .web) and IPC
(just in .krb). This needs some more thought so that it polishes a bit
better.

I detected that the accept_ctx function was producing output even in the
success-case, and reworked the HTTP pathway innards so we can return
this to the browser. Firefox doesn't like it (up it), but Chromium
handles it; Wireshark on the other hand does point out that Chromium's
_subsequent_ reply with an _empty_ Negotiate header is "illegal".
Captured Wireshark text-tree is in Notes-2.txt.