krb5: fixed message-sequencing with URL params
This commit updates the HTTP request handing so that we can rely on
GSSAPI tokens being passed in-order between the browser and server. We
set a URL parameter via a 302 redirect which identifies the request
sequence, and the browser faithfully passes this on all subsequent
Negotiate requests. We store this internally against the most recent
context-export, and resume from where we left off when we see the next
request.

This is really, really rough at the moment but I'm versioning it as a
working implementation before refacting.