ipc++: create project io/ from KdbIoDefs and CircularBuffer
This commit migrates the `KdbIoDefs.h` and `CircularBuffer` header and
source files into a new top-level project `io`, and updates references
and build files accordingly. The `demo_circular_buffer.cpp` file still
resides in `ipc++`, since it needs a lot of `KdbType` class definitions
in order to build the payload.

Note the inclusion of the new project `iou`, in which I show that
`io_uring` can be used with the `CircularBuffer` to write from the
join-point in mapped regions.